+++ pysgrs/__init__.py
@@ -3,7 +3,7 @@ import warnings
 from pysgrs.interfaces import *
 from pysgrs.cypher import *
 
-__version__ = "0.0.6"
+__version__ = "0.0.7"
 
 warnings.simplefilter('always')