+++ pysgrs/tests/test_substitution.py
@@ -2,7 +2,7 @@ import sys
 import unittest
 
 from pysgrs.tests.test_cypher import TestCypher
-from pysgrs.cypher import RotationCypher, CaesarCypher, ReversedCypher, PermutationCypher
+from pysgrs.cypher import RotationCypher, CaesarCypher, ReversedCypher, PermutationCypher, AffineCypher
 from pysgrs import errors
 from pysgrs import settings
 
@@ -62,9 +62,7 @@ class TestReversedCypher(TestCypher, unittest.TestCase):
 class TestPermutationIdentityCypher(TestCypher, unittest.TestCase):
 
     cypher = PermutationCypher()
-    cyphers = [
-        "ABCDEFGHIJKLMNOPQRSTUVWXYZ",
-    ]
+    cyphers = TestCypher.sentences
 
 
 class TestPermutationCypher(TestCypher, unittest.TestCase):
@@ -74,6 +72,14 @@ class TestPermutationCypher(TestCypher, unittest.TestCase):
                                 19, 14, 16, 11,  7,  4, 23, 17])
     cyphers = [
         "KYISPNBZJWUGCAFDMVTOQLHEXR"
+    ]
+
+
+class TestAffineCypher(TestCypher, unittest.TestCase):
+
+    cypher = AffineCypher()
+    cyphers = [
+        "INSXCHMRWBGLQVAFKPUZEJOTYD"
     ]